+++ Junk/sqlar.c
